adjust base entities
