+++ homeassistant/components/husqvarna_automower/lawn_mower.py
@@ -23,7 +23,7 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from . import AutomowerConfigEntry
 from .const import DOMAIN
 from .coordinator import AutomowerDataUpdateCoordinator
-from .entity import AutomowerControlEntity
+from .entity import AutomowerAvailableEntity
 
 DOCKED_ACTIVITIES = (MowerActivities.PARKED_IN_CS, MowerActivities.CHARGING)
 MOWING_ACTIVITIES = (
@@ -94,7 +94,7 @@ async def async_setup_entry(
     )
 
 
-class AutomowerLawnMowerEntity(AutomowerControlEntity, LawnMowerEntity):
+class AutomowerLawnMowerEntity(AutomowerAvailableEntity, LawnMowerEntity):
     """Defining each mower Entity."""
 
     _attr_name = None
